# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger LawyerUp - OpenAPI 3.0",
    "description": "This is a LawyerUp Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "email": "taha.muhammad@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "For registering a user using phone number",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "password",
                  "phone",
                  "userType"
                ],
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "string"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "string"
                  },
                  "email": {
                    "type": "string",
                    "example": "string"
                  },
                  "password": {
                    "type": "string",
                    "example": "string"
                  },
                  "phone": {
                    "type": "string",
                    "example": "string"
                  },
                  "userType": {
                    "type": "string",
                    "example": "INDIVIDUAL"
                  },
                  "workInfo": {
                    "type": "array",
                    "example": [
                      {
                        "companyName": "string",
                        "jobTitle": "string",
                        "startDate": "dateString",
                        "endDate": "dateString",
                        "description": "string",
                        "business": "busines_uid"
                      }
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "On success it will return 3 key value pairs (1) token: an Api access token (2) user: an object containing the details of the user (3) newlyCreated: to verify that the user was existed before this request or it is newly created in this request"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "For login with email and password.",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "string"
                  },
                  "password": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Api will return an access token and a user object."
          },
          "403": {
            "description": "Invalid email or password OR Your account is suspended OR Too many invalid login attempts, your login with email has been locked temporarily, please login with your phone to continue "
          }
        }
      }
    },
    "/auth/send-email-code": {
      "post": {
        "tags": ["Auth - Email Code"],
        "summary": "For sending email code.",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "source"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "string"
                  },
                  "source": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/auth/verify-email-code": {
      "post": {
        "tags": ["Auth - Email Code"],
        "summary": "For verifying email code.",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "source", "verificationToken"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "string"
                  },
                  "source": {
                    "type": "string",
                    "example": "EMAIL_VERIFICATION"
                  },
                  "verificationToken": {
                    "type": "number",
                    "example": "123456"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Failure"
          }
        }
      }
    },
    "/auth/update-password": {
      "post": {
        "tags": ["Auth - Update Password"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For updating your password.",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": {
                  "password": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Failure"
          },
          "401": {
            "description": "Failure"
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "tags": ["Search"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For searching by category and state.",
        "description": "",
        "parameters": [
          {
            "name": "searchId",
            "in": "query",
            "description": "",
            "required": false,
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "description": "",
            "required": false,
            "type": "string"
          },
          {
            "name": "state",
            "in": "query",
            "description": "",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/search/{searchId}": {
      "get": {
        "tags": ["Search"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For getting search by searchId.",
        "description": "",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": ["Search"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For deleting search by searchId.",
        "description": "",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/search/{searchId}/bookmark": {
      "post": {
        "tags": ["Bookmark"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For adding a bookmark.",
        "description": "",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "string"
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["providerId"],
                "properties": {
                  "providerId": {
                    "type": "string",
                    "example": "provider_id"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": ["Bookmark"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For deleting a bookmark.",
        "description": "",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "string"
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["providerId"],
                "properties": {
                  "providerId": {
                    "type": "string",
                    "example": "provider_id"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/languages": {
      "get": {
        "tags": ["Language"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For getting languages.",
        "description": "No input required for this api, just needed to be authenticated.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": ["Language"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For posting a language.",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/states": {
      "get": {
        "tags": ["State"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For getting states.",
        "description": "No input required for this api, just needed to be authenticated.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": ["State"],
        "security": [{ "bearerAuth": [] }],
        "summary": "For posting a State.",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "code"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "string"
                  },
                  "code": {
                    "type": "string",
                    "example": "lang_code"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}